Updated plotting mechanism radically.
Plotting script is now more flexible and makes it possible to do more
elaborate setup of each individual plot. This is done by adding a JSON
file that holds a record for each plot in the documentation. A plot
can now be restricted to a given bounding box, and there's a choice
between plotting as polygons or linestrings, as well as both low and
medium resolution data to go choose from.

More or less every projection plot in the documentation has been
updated in this commit, to benefit from the better plot options. Most
plots are now done as polygons. Generally speaking the plots that are
still based on line data have show some problems when plotted as
polygons. There is still an unsolved wrap-around problem that should
eventually be fixed.
The land.geosjson dataset has been densified manually at the antartic
to make sure that the land mass is depicted correctly in projected
space.